<ADD> validation of request body

+++ backend/routes/items.js
@@ -4,22 +4,60 @@ const router = express.Router();
 const app = express();
 
 router.get("/", async (req,res, next)=>{
-    await listItems(req,res)
+    try{
+        await listItems(req,res)
+    }
+    catch(e){
+        console.log(e)
+    }
 })
 
 //add
 router.post("/", async (req,res,next)=>{
-    await addItem(req,res)
+    if (Object.keys(req.body).length === 0) {
+        // Request body is empty or not present
+        res.status(400).send("Request body is required");
+    } else {
+        try{
+            await addItem(req,res)
+        }
+        catch(e){
+            console.log(e)
+        
+        }
+    }
 })
 
 //update
 router.put("/", async(req,res,next)=>{
-    await editItem(req,res)
+    if (Object.keys(req.body).length === 0) {
+        // Request body is empty or not present
+        res.status(400).send("Request body is required");
+    } 
+    else{
+        try{
+            await editItem(req,res)
+        }
+        catch(e){
+            console.log(e)
+        }
+    }
 })
 
 //delete
 router.delete("/", async(req,res,next)=>{
-    await deleteItem(req,res)
+    if (Object.keys(req.body).length === 0) {
+        // Request body is empty or not present
+        res.status(400).send("Request body is required");
+    } 
+    else{
+        try{
+            await deleteItem(req,res)
+        }
+        catch(e){
+            console.log(e)
+        }
+    }
 })
